refactor:DogAnimationのAnimateとUnhappyOrHungryWalkinganimateを分離
